Разделяет js на модули
